chore: move files to correspondent packages
