+++ android/src/main/java/com/adyenreactnativesdk/action/ActionComponentDialogFragment.kt
@@ -4,7 +4,7 @@
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  */
 
-package com.adyenreactnativesdk
+package com.adyenreactnativesdk.action
 
 import android.content.DialogInterface
 import android.os.Bundle
@@ -20,6 +20,9 @@ import com.adyen.checkout.core.exception.ComponentException
 import com.adyen.checkout.core.log.LogUtil
 import com.adyen.checkout.core.log.Logger
 import com.adyen.checkout.dropin.databinding.FragmentActionComponentBinding
+import com.adyenreactnativesdk.ActionHandler
+import com.adyenreactnativesdk.ActionHandlerConfiguration
+import com.adyenreactnativesdk.ActionHandlingInterface
 import com.google.android.material.bottomsheet.BottomSheetDialogFragment
 
 class ActionComponentDialogFragment(
