+++ android/src/main/java/com/adyenreactnativesdk/component/dropin/DropInServiceProxy.java
@@ -4,7 +4,7 @@
  * This file is open source and available under the MIT license. See the LICENSE file for more info.
  */
 
-package com.adyenreactnativesdk;
+package com.adyenreactnativesdk.component.dropin;
 
 import androidx.annotation.NonNull;
 import androidx.annotation.Nullable;
