+++ android/src/main/java/com/adyenreactnativesdk/ui/PaymentComponentListener.java
@@ -1,9 +1,8 @@
-package com.adyenreactnativesdk;
+package com.adyenreactnativesdk.ui;
 
 import com.adyen.checkout.components.model.payments.request.PaymentComponentData;
 
 public interface PaymentComponentListener {
     void onError(Exception exception);
-
     void onSubmit(PaymentComponentData data);
 }
